# docs/spec/api-openapi.yaml
openapi: 3.0.3
info:
  title: SEI Manager API
  version: 0.1.0
servers:
  - url: https://localhost:3000
paths:
...
  /api/processes:
    get:
      summary: Listar processos
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: start
          schema: { type: string, format: date }
        - in: query
          name: end
          schema: { type: string, format: date }
      responses:
        '200': { description: OK }
    post:
      summary: Criar processo
      responses:
        '201': { description: Created }
...
  /api/report:
    get:
      summary: Relatório unificado por período
      parameters:
        - in: query
          name: start
          required: true
          schema: { type: string, format: date }
        - in: query
          name: end
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: OK }
...